Initial I/O vector implementation
Largely untested implementation of IO vectors, as suggested in #17. Example definitions updated to use this.

Network inputs can be defined as an n-length vector like so:
    inputs : {
        vec[8] : false
    }

Network outputs can be defined as an n-length vector like so:
    outputs : {
        vec[4] : {
            0 : dest.pin,
            1 : dest.pin,
            2 : dest.pin,
            3 : dest.pin
        }
    }

Two vectors of the same length can be connected together in a connections block like so:
    vec1[] : vec2[]

Individual elements of a vector can be accessed like C style arrays, vec[2] references the third element in the array.
No support currently for connecting a vector network output to a vector of a component, must define each connection individually as in the above example.