More complex example defs
Some scripts to generate arbitrary width synchronous counters, ripple adders and carry-lookahead adders.
Note that the filesizes for the synchronous counter lookahead adder gets very large with large sizes (500kb for a 64-bit carry-lookahead adder), so minifying the files is a good idea. The last couple of lines of the scripts support this already.